+++ src/QueryType/Server/Configsets/Result/ListConfigsetsResult.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the Solarium package.
+ *
+ * For the full copyright and license information, please view the COPYING
+ * file that was distributed with this source code.
+ */
+
+namespace Solarium\QueryType\Server\Configsets\Result;
+
+use Solarium\QueryType\Server\Query\AbstractResult;
+
+/**
+ * ConfigsetsResult.
+ */
+class ListConfigsetsResult extends AbstractResult
+{
+    /**
+     * @return string[]
+     */
+    public function getConfigSets(): array
+    {
+        $this->parseResponse();
+
+        return $this->data['configSets'] ?? [];
+    }
+}