+++ tests/Integration/SolrCloud/Psr18Test.php
@@ -7,7 +7,6 @@ use Solarium\Tests\Integration\AbstractCloudTest;
 /**
  * @group integration
  * @group skip_for_solr_server
- * @coversNothing
  */
 class Psr18Test extends AbstractCloudTest
 {