+++ tests/Integration/SolrServer/Psr18Test.php
@@ -7,7 +7,6 @@ use Solarium\Tests\Integration\AbstractServerTest;
 /**
  * @group integration
  * @group skip_for_solr_cloud
- * @coversNothing
  */
 class Psr18Test extends AbstractServerTest
 {